Feature: handle rescue_from Module arguments properly
There is a common pattern for `rescue_from` and alike methods usage, to
pass as an argument a module which handles similar exceptions
processing. The most useful case (and the reason why I proposed that
change) is to handle different network errors with only on class.

In our project it's called Handle::NetworkErrors, it's a module which
aggregates handling of more then 20 different errors whose meaning is
just "something was wrong with the network" and in 99% of the cases it's
reasonable to handle them in exact similar way.

Unfortunately, without proposed change it's impossible to pass to
graphql-errors `rescue_from` a module. It has to be a class...
The other implementations of same method in Rails and Sidekiq accepts
Module as well as a Class.

So, my point is that would be great to have `rescue_from` behavior in a
cannonical way (or you may call it, principle of least surprise).